+++ content/units/vietnam/copter-c.hjson
@@ -21,6 +21,7 @@ weapons: [
         y: 14
         name: copter-c-gun
         reload: 20
+        shootCone: 10
         range: 180
         layerOffset: -1
         bullet: {
@@ -48,6 +49,7 @@ weapons: [
         y: 10
         range: 180
         reload: 60
+        shootCone: 10
         bullet: {
             maxRange: 250
             drag: -0.02
