+++ scripts/main.js
@@ -1,3 +1,4 @@
 require("units/weapons");
 require("units/units");
+require("units/vietnam/copters");
 require("blocks/units/constructor-pad")
