+++ scripts/units/units.js
@@ -10,25 +10,7 @@ scudLauncher.weapons.add(weapons.scudWeapon);
 //harvester
 const dongBaoA = extend(UnitType, "harvester", {});
 dongBaoA.constructor = () => extend(UnitEntity, {});
-//copter t1
-const dongBaoB = extend(UnitType, "copter-c", {
-    draw(unit){
-        this.super$draw(unit)
-        Draw.rect(
-            Core.atlas.find("commandustry-copter-c-blade"),
-            unit.x + Angles.trnsx(unit.rotation, 6),
-            unit.y + Angles.trnsy(unit.rotation, 6),
-            Time.time * 15
-        )
-        Draw.rect(
-            Core.atlas.find("commandustry-copter-c-top"),
-            unit.x + Angles.trnsx(unit.rotation, 6),
-            unit.y + Angles.trnsy(unit.rotation, 6),
-            unit.rotation - 90
-        )
-    }
-});
-dongBaoB.constructor = () => extend(UnitEntity, {});
+
 //export for unit pad script
 module.exports = {
     dongBaoA:dongBaoA
